refactor(models): lift the hangar's ship picker out into a shared modal
The filterable card grid the hangar adds ships through lived entirely inside
NewVehiclesModal, tied to the vehicle bulk-create mutation. Everything above that
mutation - the search, the manufacturer and classification filters, the paged
grid, the selection tray - is a ship picker, not a hangar feature.

So it moves to Models/PickerModal, which emits its selection and knows nothing
about what the caller does with it, and NewVehiclesModal becomes the wrapper that
holds the title, the submit label and the mutation. Hangar and wishlist behave
exactly as before.

Three props generalise what the hangar had hardcoded: `quantities` gates the
stepper, since only a list that can hold the same ship twice has use for it;
`max`/`maxHint` cap a selection; `takenSlugs`/`takenNote` show what the caller's
list already holds without letting it be picked again.

The translations move with it - the `newVehicles` namespace becomes
`modelPicker`, and `title` becomes `hangarTitle` beside the wishlist one.
